requirements.txt
pandas
numpy
scikit-learn
matplotlib
seaborn
joblib
statsmodels
dotenv
psycopg2-binary
streamlit